Create start up sequence to cache registry names
Thie idea is that the start_operator function should get called as soon
as kopf import the handlers subpackage. Then it will search for existing
StrimziSchemaRegistry deployments, and cache those into a state module.
This will make it easier to monitor secrets later to ensure that we
watch secrets corresponding to StrimziSchemaRegistry deployments.