Convert resources to dicts
This will make them easier and more conventional to use if everything is
a dict.

The kubernetes client converts Kubernetes API fields from snake case to
camel case, which is canonical for working with Kubernetes resources.
To work around that we use either _preload_content on the return or
sanitize_for_serialization on the python-native object.