Add workflow to run Kafka and schema registry
This is initial work to create an integration test for Strimzi Registry
Operator that runs in minikube, and is packaged as a GitHub Actions
workflow.

This is based on the Strimzi video: https://youtu.be/4bKSPrENDQQ

We're using the medyagh/setup-minikube action to run minikube in GitHub
Actions since this is supported by the minikube developers.

All the Kubernetes resources and deployment scripts are located in
minikube/

- The setupkafka.sh script deploys a Kafka cluster with Strimzi
  inside minikube.
- The buildimage.sh script is responsible for building the strimzi
  registry operator image *locally* within the minikube environment, so
  that the branch can be tested without pulling from a container registry.
  This uses minikube's docker-env feature.

- The deploysro.sh script runs a deployment of the strimzi schema
  registry operator using kustomization directory, operator-deployment/
  This kustomization essentially changes the operator's deployment to pull
  from the local image in minikube.
- The deployregistry.sh script deployed the schema registry resource
  along with the strimzi KafkaUser and KafkaTopic resources required to
  run a registry.
- The testregistry.sh script tests the schema registry by exercising
  its API.